Move sitk2vtk to vtk submodule
The following import no longer works:

from SimpleITK.utilities import sitk2vtk

This needs to be changed to:

from SimpleITK.utilities.vtk import sitk2vtk

Removed the import of the large vtk package when importing
Simpleitk.utilities. This can improve performance when vtk is not
needed when using other provided utilities. This follow the pattern of
other optional dependencies for the Utilities.